+++ src/Parametizer/Script/ScriptLauncher/Subcommand/ScriptLauncherScriptAbstract.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MagicPush\CliToolkit\Parametizer\Script\ScriptLauncher\Subcommand;
+
+use MagicPush\CliToolkit\Parametizer\Script\Subcommand\ScriptAbstract;
+
+abstract class ScriptLauncherScriptAbstract extends ScriptAbstract {
+    public static function getNameSections(): array {
+        return array_merge(parent::getNameSections(), ['script-launcher']);
+    }
+}