+++ src/Parametizer/Script/Subcommand/ScriptAbstract.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace MagicPush\CliToolkit\Parametizer\Script;
+namespace MagicPush\CliToolkit\Parametizer\Script\Subcommand;
 
 use MagicPush\CliToolkit\Parametizer\CliRequest\CliRequest;
 use MagicPush\CliToolkit\Parametizer\Config\Builder\BuilderInterface;