+++ tests/Tests/Parametizer/BuiltInSubcommand/scripts/subcommands-with-name-sections.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use MagicPush\CliToolkit\Parametizer\Script\ScriptAbstract;
+use MagicPush\CliToolkit\Parametizer\Script\Subcommand\ScriptAbstract;
 use MagicPush\CliToolkit\Tests\Utils\TestUtils;
 
 require_once __DIR__ . '/../../../init-console.php';