+++ tests/Tests/Parametizer/Subcommand/scripts/same-name-different-branches.php
@@ -27,7 +27,7 @@ $request = TestUtils::newConfig()
 
     ->run();
 
-echo $request->getParam(CliRequest::SUBCOMMAND_PREFIX . $request->getSubcommandRequestName())['opt']
+echo $request->getParam(CliRequest::SUBCOMMAND_PREFIX . $request->getRequestedSubcommandName())['opt']
     . ', '
     // Let's show that the method below is a more handy version of reading nested parameters, rendering the same values.
     . $request