+++ tools/cli-toolkit/Scripts/CliToolkitScriptAbstract.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace MagicPush\CliToolkit\Tools\CliToolkit\Scripts;
 
-use MagicPush\CliToolkit\Parametizer\Script\ScriptAbstract;
+use MagicPush\CliToolkit\Parametizer\Script\Subcommand\ScriptAbstract;
 
 abstract class CliToolkitScriptAbstract extends ScriptAbstract {
     public static function getNameSections(): array {